Refactored by replacing price type with BigDecimal

+++ src/main/java/Item.java
@@ -1,10 +1,12 @@
+import java.math.BigDecimal;
 import java.text.DecimalFormat;
 
 public class Item {
-    private final double price;
+    private final BigDecimal price;
+    //private final double price;
     //private final String name;
 
-    public Item(String name, double price) {
+    public Item(String name, BigDecimal price) {
         //this.name = name;
         this.price = price;
     }
